+++ lib/src/lib/storages/storage-map.spec.ts
@@ -18,8 +18,8 @@ import { StorageMap } from "./storage-map";
 function tests(description: string, localStorageServiceFactory: () => StorageMap): void {
 
   interface Monster {
-    name: string;
-    address?: string;
+    readonly name: string;
+    readonly address?: string;
   }
 
   const key = "test";
@@ -328,7 +328,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             items: { type: "string" },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: string[] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get(key, schema))
+          ).subscribe((result: readonly string[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -343,7 +345,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             items: { type: "integer" },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: number[] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get(key, schema))
+          ).subscribe((result: readonly number[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -358,7 +362,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             items: { type: "number" },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: number[] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get(key, schema))
+          ).subscribe((result: readonly number[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -373,7 +379,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             items: { type: "boolean" },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: boolean[] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get(key, schema))
+          ).subscribe((result: readonly boolean[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -391,7 +399,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get<string[][]>(key, schema))).subscribe((result: string[][] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get<string[][]>(key, schema))
+          ).subscribe((result: readonly (readonly string[])[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -420,7 +430,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             },
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get<Monster[]>(key, schema))).subscribe((result: Monster[] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get<Monster[]>(key, schema))
+          ).subscribe((result: readonly Monster[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -437,7 +449,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             uniqueItems: true,
           } satisfies JSONSchema;
 
-          storage.set(key, Array.from(value), schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: string[] | undefined) => {
+          storage.set(key, Array.from(value), schema).pipe(
+            mergeMap(() => storage.get(key, schema)),
+          ).subscribe((result: readonly string[] | undefined) => {
 
             expect(result).toEqual(array);
 
@@ -467,13 +481,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             }],
           } satisfies JSONSchema;
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get<[
-            string,
-            Monster
-          ]>(key, schema))).subscribe((result: [
-            string,
-            Monster
-          ] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get<[string, Monster]>(key, schema))
+          ).subscribe((result: readonly [string, Monster] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -512,13 +522,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
             },
           } satisfies JSONSchema;
 
-          storage.set(key, Array.from(value), schema).pipe(mergeMap(() => storage.get<[
-            string,
-            Monster
-          ][]>(key, schema))).subscribe((result: [
-            string,
-            Monster
-          ][] | undefined) => {
+          storage.set(key, Array.from(value), schema).pipe(
+            mergeMap(() => storage.get<[string, Monster][]>(key, schema)),
+          ).subscribe((result: readonly (readonly [string, Monster])[] | undefined) => {
 
             expect(result).toEqual(array);
 
@@ -532,15 +538,15 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         it("with all subtypes", () => {
 
           interface User {
-            name: string;
-            age: number;
-            philosopher: boolean;
-            books: string[];
-            family: {
-              brothers: number;
-              sisters: number;
+            readonly name: string;
+            readonly age: number;
+            readonly philosopher: boolean;
+            readonly books: readonly string[];
+            readonly family: {
+              readonly brothers: number;
+              readonly sisters: number;
             };
-            creditCard?: number;
+            readonly creditCard?: number;
           }
 
           const value: User = {
@@ -586,8 +592,8 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         it("without required properties", () => {
 
           interface User {
-            name?: string;
-            age?: number;
+            readonly name?: string;
+            readonly age?: number;
           }
 
           const value: User = {
@@ -611,7 +617,7 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         it("objects / cast / no schema", () => {
 
           interface Test {
-            test: string;
+            readonly test: string;
           }
 
           // @ts-expect-error Failure test
@@ -623,7 +629,7 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         it("objects / no cast / schema", () => {
 
           interface Test {
-            test: string;
+            readonly test: string;
           }
 
           storage.get("test", {
@@ -678,10 +684,10 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
 
           const value = new Blob();
 
-          storage.set(key, value).pipe(mergeMap(() => storage.get(key))).subscribe((storage.backingEngine === "localStorage") ? {
-            next: (): void => {
-              // Nothing to do
-            },
+          storage.set(key, value).pipe(
+            mergeMap(() => storage.get(key))
+          ).subscribe((storage.backingEngine === "localStorage") ? {
+            next: (): void => {},
             error: (): void => {
               done();
               return;
@@ -701,15 +707,12 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         it("heavy schema", () => {
 
           interface City {
-            country: string;
-            population: number;
-            coordinates: [
-              number,
-              number
-            ];
-            monuments?: {
-              name: string;
-              constructionYear?: number;
+            readonly country: string;
+            readonly population: number;
+            readonly coordinates: readonly [number, number];
+            readonly monuments?: readonly {
+              readonly name: string;
+              readonly constructionYear?: number;
             }[];
           }
 
@@ -776,13 +779,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
           } satisfies JSONSchema;
 
 
-          storage.set(key, value, schema).pipe(mergeMap(() => storage.get<[
-            string,
-            City
-          ][]>(key, schema))).subscribe((result: [
-            string,
-            City
-          ][] | undefined) => {
+          storage.set(key, value, schema).pipe(
+            mergeMap(() => storage.get<[string, City][]>(key, schema)),
+          ).subscribe((result: readonly (readonly [string, City])[] | undefined) => {
 
             expect(result).toEqual(value);
 
@@ -1220,7 +1219,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
         const value = ["hello 1", "hello 2"];
         const schema = Type.Array(Type.String());
 
-        storage.set(key, value, schema).pipe(mergeMap(() => storage.get(key, schema))).subscribe((result: string[] | undefined) => {
+        storage.set(key, value, schema).pipe(
+          mergeMap(() => storage.get(key, schema))
+        ).subscribe((result: readonly string[] | undefined) => {
 
           expect(result).toEqual(value);
 
@@ -1246,13 +1247,9 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
           })
         ]);
 
-        storage.set(key, value, schema).pipe(mergeMap(() => storage.get<[
-          string,
-          Monster
-        ]>(key, schema))).subscribe((result: [
-          string,
-          Monster
-        ] | undefined) => {
+        storage.set(key, value, schema).pipe(
+          mergeMap(() => storage.get<[string, Monster]>(key, schema))
+        ).subscribe((result: readonly [string, Monster] | undefined) => {
 
           expect(result).toEqual(value);
 
@@ -1264,15 +1261,15 @@ function tests(description: string, localStorageServiceFactory: () => StorageMap
       it("object", () => {
 
         interface User {
-          name: string;
-          age: number;
-          philosopher: boolean;
-          books: string[];
-          family: {
-            brothers: number;
-            sisters: number;
+          readonly name: string;
+          readonly age: number;
+          readonly philosopher: boolean;
+          readonly books: readonly string[];
+          readonly family: {
+            readonly brothers: number;
+            readonly sisters: number;
           };
-          creditCard?: number;
+          readonly creditCard?: number;
         }
 
         const value: User = {