+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This library is fully documented and so you will find detailed [migration guides](https://github.com/cyrilletuzi/angular-async-local-storage/blob/main/MIGRATION.md).
 
+## 22.0.0 (2026-06-04)
+
+### Breaking changes
+
+- Angular 22 is required.
+- `StorageConfig` and `JSONSchema` properties have been marked as `readonly`. This is mainly to enforce immutability inside of the library to avoid bugs, it should not have impacts at usage.
+
 ## 21.0.0 (2025-11-20)
 
 ### Breaking changes